requirements.txt
requests>=2.31.0
python-nmap>=0.7.1
builtwith>=1.3.4
python-whois>=0.8.0
